changed all exceptions raised from Exception to AssertionError. This avoids obscure errors when exceptions are raised while running under joblib.
This is because joblib tries to wrap "non standard exceptions" with a class it creates on the fly to inherit from both the original exception and JoblibException.
This in turn fails with the message "Cannot create a consistent method resolution" when the original exception is Exception.
See here: http://stackoverflow.com/questions/3003053/metaclass-multiple-inheritance-inconsistency